Convert makeflat tests to gtest
