+++ changes/+makeflat-gtest-conversion.change.md
@@ -1,0 +1,1 @@
+Migrated `makeflat` legacy Makefile tests to gtest/CTest. The converted tests retain the existing external ISISTESTDATA fixtures (87 MB).
