+++ isis/src/base/apps/makeflat/makeflat.h
@@ -1,0 +1,10 @@
+#ifndef makeflat_h
+#define makeflat_h
+
+#include "UserInterface.h"
+
+namespace Isis {
+  extern void makeflat(UserInterface &ui);
+}
+
+#endif
